requirements.txt
flask
flask-cors
tensorflow-cpu
pillow
numpy
opencv-python
gunicorn
gdown